lti_synchronization/requirements.txt
josepy
jupyterhub
jwcrypto
loguru
nbgrader
oauthenticator
oauthlib
pem
pycryptodome
pyjwt==1.7.1
pytest-asyncio
pytest-clarity
requests
